Setting for overriding the localhost-only option.
