+++ resources/IBController.ini
@@ -207,6 +207,13 @@ ShowAllTrades=no
 ForceTwsApiPort=
 
 
+# Override the setting "Allow connections from localhost only". If
+# AllowExternalIps is set to true, connections from all IPs configured
+# in jts.ini will be possible. If set to false, only connections from
+# localhost will be possible. If not set, no change will be made.
+
+AllowExternalIps=
+
 # If ReadOnlyLogin is set to 'yes', and the user is enrolled in IB's
 # account security programme, the user will not be asked to supply 
 # the security code, and login to TWS will occur automatically in
